MultiAgent Interview Bot/pyproject.toml
[project]
name = "multiagent-interview-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "gtts>=2.5.4",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.6",
    "langchain-groq>=0.3.5",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.1",
    "pyaudio>=0.2.14",
    "pypdf>=5.7.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "speechrecognition>=3.14.3",
    "streamlit>=1.47.0",
]
